+++ stencil-workspace/src/components/modus-time-input/modus-time-picker.types.tsx
@@ -1,0 +1,10 @@
+export type TimeParts = {
+  hours: number;
+  minutes: number;
+  ampm: string;
+};
+
+export type TimeInputEventData = {
+  value: string;
+  inputString: string;
+};