+++ CHANGELOG.md
@@ -6,6 +6,10 @@ View all releases at: <https://github.com/trimble-oss/modus-web-components/relea
 
 ## Unreleased
 
+### Added
+
+- Added Modus Time Picker component
+
 ## 0.1.33 - 2021-02-16
 
 ### Fixed